Rename files and components to be consistent
