+++ expense-splitter/src/components/friend/Friend.jsx
@@ -1,15 +1,15 @@
-import { UseDataContext } from "../components/context/SiteContext";
-import Button from "../components/ui/Button";
-import FriendsList from "./FriendsList";
+import { UseDataContext } from "../context/SiteContext";
+import Button from "../ui/Button";
+import FriendList from "./FriendList";
 
-const FriendsView = () => {
+const Friend = () => {
   const { modal, handleSetModal } = UseDataContext();
 
   return (
     !modal.show && (
       <>
         <h1 className="text-center">Friends</h1>
-        <FriendsList />
+        <FriendList />
         <Button
           className="over absolute bottom-6 left-1/2 z-10 h-14 w-[200px] -translate-x-1/2 rounded-md bg-primary text-lg"
           onClick={() => {
@@ -23,4 +23,4 @@ const FriendsView = () => {
   );
 };
 
-export default FriendsView;
+export default Friend;
