+++ expense-splitter/src/components/ui/Modal.jsx
@@ -2,7 +2,7 @@ import EditGroup from "../group/EditGroup";
 import CreateGroup from "../group/CreateGroup";
 import { UseDataContext } from "../context/SiteContext";
 import CreateExpense from "../expense/CreateExpense";
-import CreateFriend from "../../friends/CreateFriend";
+import CreateFriend from "../friend/CreateFriend";
 
 // map component names to actual components
 const components = {
